change data type storing storage amounts to bigintegerfield

+++ physionet-django/project/models.py
@@ -872,8 +872,8 @@ class PublishedProject(Metadata, SubmissionInfo):
 
     """
     # File storage sizes in bytes
-    main_storage_size = models.PositiveIntegerField(default=0)
-    compressed_storage_size = models.PositiveIntegerField(default=0)
+    main_storage_size = models.BigIntegerField(default=0)
+    compressed_storage_size = models.BigIntegerField(default=0)
     publish_datetime = models.DateTimeField(auto_now_add=True)
     is_newest_version = models.BooleanField(default=True)
     newest_version = models.ForeignKey('project.PublishedProject', null=True,
